Big Changes in Strucure of this thing...
First of all: File concurrency shouldn't be an issue now since there's
barely any config file now and all that stuff are moved into Redis db now.
Then there's Celery, I added it in because of the previous thing I mention.
It serves as configurations' task manager now so if flask want those
jsons, it will be come through Celery.
I also moved generate_working_server() to Celery as update_working_server()
and modified it to be able to interact with Redis.

That's about this commit. I'll spare the details because that's a lot.